+++ esphome/components/mcp2515/mcp2515.cpp
@@ -113,7 +113,7 @@ uint8_t MCP2515::get_status_() {
 canbus::Error MCP2515::set_mode_(const CanctrlReqopMode mode) {
   modify_register_(MCP_CANCTRL, CANCTRL_REQOP, mode);
 
-  unsigned long end_time = millis() + 10;
+  uint32_t end_time = millis() + 10;
   bool mode_match = false;
   while (millis() < end_time) {
     uint8_t new_mode = read_register_(MCP_CANSTAT);
@@ -600,9 +600,9 @@ canbus::Error MCP2515::set_bitrate_(canbus::CanSpeed can_speed, CanClock can_clo
   }
 
   if (set) {
-    set_register_(MCP_CNF1, cfg1);
-    set_register_(MCP_CNF2, cfg2);
-    set_register_(MCP_CNF3, cfg3);
+    set_register_(MCP_CNF1, cfg1);  // NOLINT
+    set_register_(MCP_CNF2, cfg2);  // NOLINT
+    set_register_(MCP_CNF3, cfg3);  // NOLINT
     return canbus::ERROR_OK;
   } else {
     return canbus::ERROR_FAIL;
